Add Dockerfile for server
This Dockerfile builds an image for running the server. A multi-stage
build is used where the first stage installs the Python bits, and the
second stage grabs all of /usr/local and dumps that on top of the stock
official node image. After a dynamic linker cache rebuild, everything
seems to work just fine.